PLAT-9093: Build and release with Github actions
Switch to Github actions (even better integration with Github) to build
and release the project.

Build is enabled for PR and master branch, mostly running the unit
tests.

A release workflow is also defined, to be executed manually. It uses the
Maven release plugin and will also create the Github release from the
tag. It takes the version numbers as input. What will be left to do once
released is to complete the Github release description.

To be able to release, secrets are defined at the project level (GPG key
and Sonatype credentials).

Also update badges in README.