MML-258 Set up framewok to start collecting BI events
Goal of this PR is to set up a framework in order to be able collect BI events.
The BiContext is the main object which contains a list of BiItems (one per element found inside the messageML). While parsing the MessageML the BiContext will be initialized and automaticallty filled with all the information.
Each element will create a new BiItem which is composed by the element name and a map of all attributes found. To customize this item each element can override the updateBiContext() method of Element class and define the map as needed. For now the function is by default creating a simple item which counts the occurrences of the element inside the message. An example of custom biItem can be found in the Button class.