MML-265 Fill BiContext with MessageML simple tags attributes
After this implementation #259 we are ready to start filling the BiContext with a list BiItems.
This PR  will take care of all the simple tags elements such as headers, paragraphs, link. Most of the cases for these elements we want to compute the number of occurrences of the elements it self or in some cases the max value found inside the message (like for TableColumnsMax).
Also we are going to add general BI to compute MessageLength, FreeMarkerUse, EntityJSONSize and usage of Styles and Custom css.

Each one of this BiItem is following the same format. Since for them we are only going to store a single attribute in the map, each item has a specific name and a map with only "count" attribute.