Compute presentationML only once
When used by the Agent it ends up being computed twice, we can cache it
to avoid that.

This is not thread safe but the worst that could happen is that it will
be computed twice so we should be good.

From 9000 ops/s to a bit over 9000 ops/s

https://perzoinc.atlassian.net/wiki/spaces/DevX/pages/2106428894/Agent+CPU+usage+investigations