Error handling for EntityJson parsing
If a string (i.e valid JSON) is provided as data (invalid EntityJSON)
a ClassCastException was raised.
We try to parse it as an escaped JSON string if possible otherwise
this is now handled as an InvalidInputException with a better error message.

This allow clients (BDKs) to send EntityJSON as data that is serialized
to JSON by the BDK or as raw JSON strings.

Fixes PLAT-11318