PLAT-11467 Update BiFields to support default value
The goal here is to be able to differentiate 3 things. The element name, Bi properties that should be strings and Bi properties that should be integers.

By setting the default value of the element name and Bi properties that should be strings to null we can make that distinction in the agent. Enabling us to create a default payload with 0 for all fields that would have an Integer as a default value and add the others if present.